--- a/activerecord/lib/active_record/migration.rb
+++ b/activerecord/lib/active_record/migration.rb
@@ -544,21 +544,38 @@ module ActiveRecord
     # This class is used to verify that all migrations have been run before
     # loading a web page if <tt>config.active_record.migration_error</tt> is set to :page_load
     class CheckPending
-      def initialize(app)
+      def initialize(app, file_watcher: ActiveSupport::FileUpdateChecker)
         @app = app
-        @last_check = 0
+        @needs_check = true
+
+        @mutex = Mutex.new
+        @watcher = file_watcher.new(*watchable_args) do
+          @needs_check = true
+          ActiveRecord::Migration.check_pending!(connection)
+          @needs_check = false
+        end
       end
 
       def call(env)
-        mtime = ActiveRecord::Migrator.last_migration.mtime.to_i
-        if @last_check < mtime
-          ActiveRecord::Migration.check_pending!(connection)
-          @last_check = mtime
+        @mutex.synchronize do
+          if @needs_check
+            @watcher.execute
+          else
+            @watcher.execute_if_updated
+          end
         end
+
         @app.call(env)
       end
 
       private
+
+        def watchable_args
+          dirs_hash = ActiveRecord::Migrator.migrations_paths.map do |dir|
+            [dir, ["rb"]]
+          end.to_h
+          [[], dirs_hash]
+        end
 
         def connection
           ActiveRecord::Base.connection
@@ -950,10 +967,6 @@ module ActiveRecord
       File.basename(filename)
     end
 
-    def mtime
-      File.mtime filename
-    end
-
     delegate :migrate, :announce, :write, :disable_ddl_transaction, to: :migration
 
     private
@@ -966,16 +979,6 @@ module ActiveRecord
         require(File.expand_path(filename))
         name.constantize.new(name, version)
       end
-  end
-
-  class NullMigration < MigrationProxy #:nodoc:
-    def initialize
-      super(nil, 0, nil, nil)
-    end
-
-    def mtime
-      0
-    end
   end
 
   class Migrator#:nodoc:
@@ -1049,10 +1052,6 @@ module ActiveRecord
 
       def any_migrations?
         migrations(migrations_paths).any?
-      end
-
-      def last_migration #:nodoc:
-        migrations(migrations_paths).last || NullMigration.new
       end
 
       def migrations_paths