`Migration::CheckPending` is a rack middleware normally used in development to raise an exception when there are pending migrations.

This commit replaces the existing caching, which avoided checking all migrations if the newest migration hasn't changed. Instead, we now use `FileUpdateChecker`, which has two advantages: it can detect new migrations which aren't the highest version number, and it is faster.

## Improved checking

The existing code watched for the value of `last_migration.mtime` to change, the modification time on the  file with the highest version number. This works most of the time, but when merging/rebasing/checking out other branches, it's likely to have "new" migrations arrive with a lower version number than the previous highest. In this event the `CheckPending` will not notice and not re-check that migrations are up to date.

What we really want is to check for the greatest `mtime` of all migrations, and re-check if any files are new. This is exactly what `ActiveSupport::FileUpdateChecker` does for us! :tada: 

## Improved performance

This middleware is probably only run in development, which we don't heavily optimize for, but this check was actually relatively slow, I started looking into this because it was taking ~10ms on each request in my development server (not reported in logs because this middleware is before dispatch).

This slowness is because `last_migration` is `migrations.last`, and `migrations` for each migration file needs to run `File.basename`, a regex to extract version/name/scope, and `camelize`.

My first attempt to improve this was to avoid the call to `camelize`, which is faster, but still not as fast as just using `FileUpdateChecker` (which surprised me, I expected a `stat` syscall for each file to be slow). This will use `EventedFileUpdateChecker` if it is set as `config.file_watcher` (it is by default), which is much faster.

## Performance testing

For testing, I made a new rails application, and generated 500 migrations

```
$ bundle exec rails new --dev sandbox
$ cd sandbox && rails c
> require 'rails/generators/active_record/migration/migration_generator'
> (1...500).each {|i| ActiveRecord::Generators::MigrationGenerator.start(["TestSomeMigrations#{i}"]) }
$ rake db:migrate
```

I have the following benchmark script
``` ruby
require 'benchmark/ips'
require './config/environment'

check_pending = ActiveRecord::Migration::CheckPending.new(->(env){})
# check_pending = ActiveRecord::Migration::CheckPending.new(->(env){}, file_watcher: ActiveSupport::EventedFileUpdateChecker)

env = {}

Benchmark.ips do |x|
  x.config(:time => 5, :warmup => 2)

  x.report("CheckPending#call") { check_pending.call(env) }
end
```

**Original**
```
CheckPending#call    111.225  (± 0.9%) i/s -    560.000  in   5.035700s
```

**This PR** (when `file_watcher == FileUpdateChecker`)
```
CheckPending#call    615.889  (± 1.9%) i/s -      3.132k in   5.087336s
```

**This PR** (when `file_watcher == EventedFileUpdateChecker`)

```
CheckPending#call    942.572k (± 0.2%) i/s -      4.727M in   5.015013s
```

~~It would be awesome to use `ActiveSupport::EventedFileUpdateChecker`, if available, which is over 1000x faster than the non evented version in my benchmark. I couldn't find a way to properly get access to 
`Rails.application.config.file_watcher` from either `CheckPending` itself or [where it is initialized in the ActiveRecord railtie](https://github.com/rails/rails/blob/master/activerecord/lib/active_record/railtie.rb#L76-L77). Any suggestions?~~ This now uses `ActiveSupport::EventedFileUpdateChecker`, if configured as `config.file_watcher`


Thanks for reading