@matthewd 2017-07-12
This is great! 😍 

> still not as fast as just using FileUpdateChecker (which surprised me, I expected a stat syscall for each file to be slow)

It is.. but the `Dir[]` inside `migrations` would be doing that anyway.

> I couldn't find a way to properly get access to Rails.application.config.file_watcher from either CheckPending itself or where it is initialized in the ActiveRecord railtie.

It should just be `config.file_watcher` inside the railtie initializer.