@jhawthorn 2017-07-12
> It should just be config.file_watcher inside the railtie initializer.

That's what I tried, but `file_watcher` is stored [on the instance of the application config](https://github.com/rails/rails/blob/58f10a31b37e9bb6e975a71aa63744f318ee043d/railties/lib/rails/application/configuration.rb#L43), unlike some other preferences which are shared via [class variables](https://github.com/rails/rails/blob/58f10a31b37e9bb6e975a71aa63744f318ee043d/railties/lib/rails/railtie/configuration.rb#L89-L96).

From a binding.pry in the railtie:

```
[1] pry(#<ActiveRecord::Railtie>)> config
=> #<Rails::Railtie::Configuration:0x0055a4e1f6fa50>
[2] pry(#<ActiveRecord::Railtie>)> Rails.application.config
=> #<Rails::Application::Configuration:0x0055a4e25a81d8
[3] pry(#<ActiveRecord::Railtie>)> Rails.application.config.file_watcher
=> ActiveSupport::EventedFileUpdateChecker
[4] pry(#<ActiveRecord::Railtie>)> config.file_watcher
NoMethodError: undefined method `file_watcher' for #<Rails::Railtie::Configuration:0x0055a4e1f6fa50>
from /home/jhawthorn/src/rails/railties/lib/rails/railtie/configuration.rb:95:in `method_missing'
```