@matthewd 2017-07-12
Ah right, wrong `config` sorry. Does https://github.com/rails/rails/blob/58f10a31b37e9bb6e975a71aa63744f318ee043d/activerecord/lib/active_record/railtie.rb#L109-L111 help?