@jhawthorn 2017-07-12
Yes! That's it!

Updated to use the application's setting for `file_watcher`