@jhawthorn 2017-08-02
I rebased (mostly to rerun tests) and everything is green. Does this need anything else to move forward?