@matthewd 2017-08-10
The existing test coverage looks a bit sparse, which makes me nervous that we'll manage to retain the right semantics long term, even assuming everything's currently in order.

At a glance, I see only [this test](https://github.com/rails/rails/blob/3ffc77a9fb7902acacc600a897fda3107f2ff303/railties/test/application/configuration_test.rb#L127-L153) that's really exercising the behaviour, and that's not doing much to convince me it can notice when pending migrations appear, when they disappear, or even that it continues complaining after it notices the problem (the `@needs_check` bit).

That's all definitely a shortcoming of the existing tests: if they were up to scratch, I don't think there'd be any call for new tests specific to this implementation... but it still feels like something we should address. This way, we can insert the new tests first, and use them to more broadly prove what we already believe -- that the observable behaviour hasn't changed.