@jhawthorn 2017-12-08
@matthewd sorry for the long delay. I finally got around to adding some additional migrations to test this and expose some shortcomings with the prior implementation.

Hopefully they're about what you had in mind.