@jhawthorn 2018-03-19
This needs to be redone over-top of the changes to migrations in 5.2 :anguished: (they are very nice changes though)

I'll open a new PR when I get the chance to redo this.